+++ waves/_settings.py
@@ -55,11 +55,6 @@ _default_output_file_type_api = _allowable_output_file_types[0]
 _default_output_file_type_cli = _allowable_output_file_types[1]
 _hash_coordinate_key = "set_hash"
 _set_coordinate_key = "set_name"
-# VVV TODO: Remove when the deprecated set coordinate key is fully removed VVV
-# https://re-git.lanl.gov/aea/python-projects/waves/-/issues/855
-_deprecated_hash_coordinate_key = "parameter_set_hash"
-_deprecated_set_coordinate_key = "parameter_sets"
-# ^^^ TODO: Remove when the deprecated set coordinate key is fully removed ^^^
 _installed_docs_index = _project_root_abspath / "docs/index.html"
 _modsim_template_directory = _project_root_abspath / "modsim_template"
 _tutorials_directory = _project_root_abspath / "tutorials"