+++ docs/changelog.rst
@@ -16,6 +16,7 @@ Breaking changes
   :merge:`1166`). By `Kyle Brindley`_.
 - Deprecate old naming convention for project help messages in favor of ``project_help_`` naming convention
   (:issue:`862`, :merge:`1168`). By `Kyle Brindley`_.
+- Remove deprecated set and hash coordinate keys (:issue:`855`, :merge:`1169`). By `Kyle Brindley`_.
 
 ********************
 0.12.10 (2025-04-10)